Changed STructure. IT IS UNTESTED.
